+++ src/App.ts
@@ -14,9 +14,7 @@ import LogHttpRequest from './Middlewares/LogHttpRequest'
 import './Database'
 
 // Router
-import AuthRouter from './Routes/Auth'
-import DocsRouter from './Routes/Docs'
-import NoMatchRouter from './Routes/NoMatch'
+import IndexRouter from './Routes'
 
 const app = express()
 
@@ -38,10 +36,6 @@ app.use(MongoSanitize)
 app.use(LogHttpRequest)
 
 // Router
-app.use('/api/auth', AuthRouter)
-
-app.use('/api/docs', DocsRouter)
-
-app.use('*', NoMatchRouter)
+app.use('/api/', IndexRouter)
 
 export default app
