+++ src/Versions/V1/Controllers/Auth/__tests__/Login.test.ts
@@ -4,15 +4,15 @@ import request from 'supertest'
 
 import { Types } from 'mongoose'
 
-import FindByEmail from '../../../Services/User/FindByEmail'
+import FindByEmail from '../../../../../Services/User/FindByEmail'
 
-import JWTSign from '../../../Utils/Promises/JWTSign'
+import JWTSign from '../../../../../Utils/Promises/JWTSign'
 
-import App from '../../../App'
+import App from '../../../../../App'
 
 jest.mock('bcrypt')
-jest.mock('../../../Services/User/FindByEmail')
-jest.mock('../../../Utils/Promises/JWTSign')
+jest.mock('../../../../../Services/User/FindByEmail')
+jest.mock('../../../../../Utils/Promises/JWTSign')
 
 const MockedJWTSign = jest.mocked(JWTSign)
 const MockedFindByEmail = jest.mocked(FindByEmail)
