+++ src/Versions/V1/Controllers/Auth/index.ts
