+++ src/Versions/V1/Routes/index.ts
@@ -1,0 +1,9 @@
+import express from 'express'
+
+import AuthRouter from './Auth'
+
+const Router = express.Router()
+
+Router.use('/auth', AuthRouter)
+
+export default Router
