+++ src/Versions/V1/Validation/Auth/Login.ts
@@ -1,7 +1,7 @@
 import { body } from 'express-validator'
 
 // Middleware
-import CheckValidationError from '../../Middlewares/CheckValidationError'
+import CheckValidationError from '../../../../Middlewares/CheckValidationError'
 
 const Login = () => {
 	const validator = [
